requirements.txt
fastapi
uvicorn
jinja2
python-dotenv
httpx
pydantic
authlib
sqlalchemy
passlib[bcrypt]
fastapi-security
redis   
python-dotenv
Flask-Login
Flask-Dance
httpx
Flask
flask-sqlalchemy
flask-migrate